Keep lower bounds in fp32

+++ fla/models/hgrn/modeling_hgrn.py
@@ -222,7 +222,7 @@ class HGRNModel(HGRNPreTrainedModel):
         all_attns = () if output_attentions else None
 
         if self.config.use_lower_bound:
-            lower_bounds = self.lower_bounds.softmax(0)
+            lower_bounds = self.lower_bounds.softmax(0, dtype=torch.float)
             lower_bounds = lower_bounds.cumsum(0) - lower_bounds[0]
         for i, layer in enumerate(self.layers):
             if output_hidden_states:
